[2.Add the data model]
Add a shipping_cost column to the Order data model
Update the Order struct to add a shipping cost field that takes a number of type float.
Update the handle_request() function to add the shipping_cost to the order.total field.
Update the SQL statements in the microservice to save the shipping cost to the database.